requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5.3
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
idna==2.10
lxml==4.5.2
multidict==4.7.6
pycodestyle==2.6.0
PyJWT==1.7.1
requests==2.24.0
selenium==3.141.0
soupsieve==2.0.1
toml==0.10.1
ujson==3.0.0
urllib3==1.25.9
yarl==1.4.2
